+++ R/ext.R
@@ -85,7 +85,6 @@ context_menu_items.dag_extension <- function(x) {
       ),
       condition = function(board, target) target$type == "node",
       id = "create_link",
-      sidebar = "actions_sidebar",
       retarget = TRUE
     ),
     new_context_menu_entry(
@@ -145,7 +144,6 @@ context_menu_items.dag_extension <- function(x) {
       ),
       condition = function(board, target) target$type == "edge",
       id = "edit_link",
-      sidebar = "actions_sidebar",
       retarget = TRUE
     ),
     new_context_menu_entry(
@@ -168,7 +166,6 @@ context_menu_items.dag_extension <- function(x) {
       ),
       condition = function(board, target) target$type == "node",
       id = "append_block",
-      sidebar = "append_block_sidebar",
       retarget = TRUE
     ),
     new_context_menu_entry(
@@ -186,8 +183,7 @@ context_menu_items.dag_extension <- function(x) {
         input_name = "ctx_add_block"
       ),
       condition = function(board, target) target$type == "canvas",
-      id = "add_block",
-      sidebar = "add_block_sidebar"
+      id = "add_block"
     ),
     new_context_menu_entry(
       name = "Create stack",
@@ -204,8 +200,7 @@ context_menu_items.dag_extension <- function(x) {
         input_name = "ctx_create_stack"
       ),
       condition = function(board, target) target$type == "canvas",
-      id = "create_stack",
-      sidebar = "actions_sidebar"
+      id = "create_stack"
     ),
     new_context_menu_entry(
       name = "Remove stack",
@@ -246,7 +241,6 @@ context_menu_items.dag_extension <- function(x) {
       ),
       condition = function(board, target) target$type == "combo",
       id = "edit_stack",
-      sidebar = "actions_sidebar",
       retarget = TRUE
     ),
     new_context_menu_entry(
@@ -370,8 +364,7 @@ toolbar_items.dag_extension <- function(x) {
       action = update_action_trigger(
         action_name = "add_block_action",
         input_name = "tool_add_block"
-      ),
-      sidebar = "add_block_sidebar"
+      )
     ),
     new_toolbar_item(
       id = "add_stack",
@@ -387,8 +380,7 @@ toolbar_items.dag_extension <- function(x) {
       action = update_action_trigger(
         action_name = "add_stack_action",
         input_name = "tool_add_stack"
-      ),
-      sidebar = "actions_sidebar"
+      )
     ),
     new_toolbar_item(
       id = "remove_selected",